fix(stickerfy): fixed stickerfying gifs and animated png uploads not being animated on sticker uploading

+++ src/commands/stickerfy.js
@@ -39,7 +39,7 @@ const uploadSticker = async (interaction, input, name, tag) => {
 				await interaction.editReply({ embeds: [sendErrorFeedback(interaction.commandName, 'Unable to upload sticker.\nLength of gif exceeds maximum duration of 5 seconds.')] });
 				break;
 			case 'Invalid Asset':
-				await interaction.editReply({ embeds: [sendErrorFeedback(interaction.commandName, 'Currently gif-to-apng support is broken. This is a temporary error due to the latest version of Discord.JS. A proper fix that restores functionality will be released soon.')] });
+				await interaction.editReply({ embeds: [sendErrorFeedback(interaction.commandName, 'Invalid asset. Please try again with a valid png or animated png.')] });
 				break;
 			default:
 				console.error(`**Command:**\n${interaction.commandName}\n**Error Message:**\n${error.message}\n**Raw Input:**\n${interaction.options.getString('url')}\n${interaction.options.getString('name')}\n${interaction.options.getString('tag')}`);
@@ -97,26 +97,30 @@ module.exports = {
 
 			// Checks if url is animated or not; if animated treat as gif, if not treat as png
 			if (isAnimated(buffer)) {
-				await sharp(buffer, { animated: true })
-					.resize(320, 320, { fit: 'contain', background: { r: 0, g: 0, b: 0, alpha: 0 } })
-					.gif({ colours: 32, dither: 0.0 })
-					.toFile(`${path}.gif`);
+				if (imageType(buffer).ext === 'png') {
+					await uploadSticker(interaction, buffer, name, tag);
+				}
+				else {
+					await sharp(buffer, { animated: true })
+						.resize(320, 320, { fit: 'contain', background: { r: 0, g: 0, b: 0, alpha: 0 } })
+						.gif({ colours: 32, dither: 0.0 })
+						.toFile(`${path}.gif`);
+					exec(`gif2apng ${path}.gif`,
+						async (execError, stdout, stderr) => {
+							if (execError) throw execError;
+							if (stderr) console.error(`gif2apng stderr ${stderr}`);
 
-				exec(`gif2apng ${path}.gif`,
-					async (execError, stdout, stderr) => {
-						if (execError) throw execError;
-						if (stderr) console.error(`gif2apng stderr ${stderr}`);
+							await uploadSticker(interaction, `${path}.png`, name, tag).finally(() => {
+								fs.unlink(`${path}.gif`, (err) => {
+									if (err) console.error(`Unable to delete image: ${err}`);
 
-						await uploadSticker(interaction, `${path}.png`, name, tag).finally(() => {
-							fs.unlink(`${path}.gif`, (err) => {
-								if (err) console.error(`Unable to delete image: ${err}`);
-
-							});
-							fs.unlink(`${path}.png`, (err) => {
-								if (err) console.error(`Unable to delete image: ${err}`);
+								});
+								fs.unlink(`${path}.png`, (err) => {
+									if (err) console.error(`Unable to delete image: ${err}`);
+								});
 							});
 						});
-					});
+				}
 			}
 			else {
 				await sharp(buffer)
